requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
gunicorn==20.1.0
kombu==5.2.3
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.24
pymemcache==3.5.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.0
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
